remove common_dev contract dependency for proxy
