+++ protocol/contracts/proxy/Cargo.toml
@@ -27,9 +27,9 @@ path = "../../contracts/common"
 default-features = false
 features = ["ink-as-dependency"]
 
-[dev-dependencies.common_dev]
+[dev-dependencies.common]
 version = "0.3.0"
-path = "../../contracts/common_dev"
+path = "../../contracts/common"
 default-features = false
 features = ["ink-as-dependency"]
